updated_requirements.txt
amqp==5.3.1
anyascii==0.3.2
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.17
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-modelcluster==6.4
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
kombu==5.4.2
l18n==2021.3
laces==0.1.2
mysqlclient==2.2.6
openpyxl==3.1.5
pillow==11.1.0
pillow_heif==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wagtail==6.3.2
wcwidth==0.2.13
Willow==1.9.0
